tests/submission/W3C/domhtml/anchor04.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Anchor</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/the-a-element.html#dom-a-href" rel="help">
<link rel="author" title="Netscape">
</head>
<body>
<p>
<a  href="../images/yellow.png">View Submit Button</a>
</p>


<script>
var domtest = async_test("The URI of the linked resource.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vhref;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("a");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vhref = testNode.href;

      assert_uri_equals(null,null,null,"yellow.png",null,null,null,true,vhref);

    domtest.done();
}));

</script>
<p>The URI of the linked resource.
The value of attribute href of the anchor element is read and checked against the expected value.</p>
<div id="log"></div>
</body></html>
